requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
requests==2.19.1
drf-extensions==0.4.0
